Use non-reserved word for name of arguments variable
The jsdoc.js file is processed twice for some reason when executing from the jsdoc3 ant task (github.com/jannon/jsdoc3-ant-task), but the arguments variable remains the same so the second pass is missing the "dirname option that was spliced out the first time.  Trouble ensues.

Although the root cause of that is unknown, there's no reason not to just be safe and name the variable something else.  Now, during the second pass, the arguments retains it's original value.