Make the tests work under Cygwin
Under Cygwin, we need to be a bit more aggressive with replacing \ with /
to avoid \ in URIs internally. This shows up when running the tests with -T,
but not when running files directly, for reasons I don't fully understand.

This patch replaces the previous replacement of \ => / with one that happens
earlier and applies to all references, not just the URIs we construct.

Cygwin now passes jsdoc -T from any directory.