Copied documentation to docs folder
I copied my current revision of the documentation into the docs folder.
Jake still works as expected. If this is accepted, then the
documentation will accompany JSDoc3 and users will be able to refer to
the specific documentation for the specific revision of JSDoc3 that they
have. The way things work currently, users only have access to the
latest revision of the documentation which may not coincide with their
version of JSDoc3.

I don't know how to set this up to be a website on the JSDoc3 project.
There is some documentation on publishing manual pages for a project on
github, but I've never done it.
https://help.github.com/articles/creating-project-pages-manually If it
works, we could figure out a way of pushing the docs directory to the
hosted pages or something...