Created .gitattributes
The directions on github say that the .gitattributes file is where a
project specifies the line endings for files. This is really stupid,
nobody asked github to convert line endings but it does it by default.
There are a bunch of different ways you can set it, this one should
specify that all files being committed will have line endings converted
to \n and, on checkout, the line endings will be converted to \n.

I don't know why they say it that way in the docs but I'm sure we'll
find out when something else happens unexpectedly. . . Maybe it'll start
spellchecking and editing "mistakes" automatically too! . . .