template.js - cache, settings now instance fields
`cache` and `settings` are no longer private fields.
This was done so
that plugins could use this class without messing up
the final output
when template name collisions occur. i.e. each instance
has it's own
template cache. The `settings` were made specific to each
instance so
that users of this class could redefine the underscore
template
settings if they want to.