Can now override which markdown parser is used.
I'm generating gfm markdown and need to specify the gfm parser in order
to get the correct output. I've added the ability to tell `getParser()`
which parser to use so that I don't have to override the global
configuration and screw everything up. If no parser is given when
calling `getParser()` or, if the parser name is unknown, it will fall
back to looking through the global configs or using the default like it
did before.