Parser configs can be specified to getParser
Using this single utility to access the markdown parsers is great for
many many reasons.

I've added the ability to override the `conf.json` settings when calling
`getParser`. It is now possible to select a specific parser and give it
custom settings. The settings given do not overwrite the global
configuration so, any subsequent usage will still be able to expect the
settings specified in `conf.json` by default.