improved --describeTags htmlFiles
The generated html is adding unique id's to things and classes to items
that could be classified. This may change because I have yet to put the
classes and id's to work. The idea here being that I'll be able to find
individual pieces and groups of information easily using the DOM.