Plugin to automatically extract a summary, if missing, from the first sentence in the description
This plugin (tries to) extract a summary, if missing, from the first sentence in the description.

It does this by splitting at the first . that is followed by a whitespace or an HTML tag. Afterwards, it appends a ., balances any HTML tags and removes a begin-end pair of P tags.

The result is pretty useful, if the documenter has not "bothered" to write a summary and description tag.