Fix for win32
I couldn't make jsdoc work from command line on windows (32bit). So I had a look at the code and a change to getModulePath function made it work. The pathToUri is also wrong because the backslash should be replaced by forward slash (anyway is now removed from any execution path with this patch).