Fix for #640 - Allow default values for parameters to contain brackets
Because JSDoc uses brackets to delimit default values inside the @param tag,
default values containing brackets can get misparsed. We solve this by
redefining splitName() in terms of splitNameMatchingBrackets() which tries
to find the matching end bracket if the name starts with an open bracket.
Ignores brackets inside strings. If splitNameMatchingBrackets() fails to find
the closing bracket, we fall back to the original regex method.