Support an Array of tags titles in allowUnknownTags
This commit adds support for specifying an Array of tags which are
unknown to JSDoc, but allowed without error.  It provides a more
granular way to disable such errors while retaining the benefits of
catching errant tags (e.g. typos).

The intended use case is catching errant tags when using additional
tools which support tags not recognized by JSDoc (e.g. Closure Compiler
and the tags discussed in #605).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>