Create silent template
This commit creates a template named `silent` which outputs nothing at
all.  As discussed in the `README.md`, this can be useful when running
JSDoc as a linter to check for syntax errors or unknown tags.  It may
also be useful for testing or benchmarking.

The same effect can be achieved using the haruki template and
redirecting the output to /dev/null, but this has a few downsides:

* It is non-obvious.
* It is difficult to do in scripts which must run in bash and cmd.exe
  (like npm scripts).
* It is slower due to the unnecessary formatting work and output
  syscalls.

This template could be distributed as a third-party template, but given
the tiny code size, broad applicability, and potential usefulness for
tests, it makes sense to include it officially.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>