Fixes directory name clashes with JSDoc plugin and template loading
The issue has been reported in #1081 (and its duplicates #1097 and #1114).
A fix was initially proposed by @GerHobbelt and merged in Dec 2015
(6be9dac616549ce226e4c0064876c7cb3b64fe25), only to be reverted in Sep
2016 (see [comment] and #1259 for why). Indeed, the initially proposed fix
was in some ways a hack, because it made assumptions that are in fact
not met in common and important situtations.

This solution takes a different route. In essence, it considers the current
behavior of jsdoc/path.getResourcePath() broken, because it did not
differentiate between filepath specifications given relative to where
packages are installed, and those relative to the current or module
directory (i.e., starting with './' or '../'). This is in sharp contrast
to how module identifiers are resolved by require(), resulting in behavior
for loading plugins and templates that differed from that of loading modules,
even though 3rd party plugins and templates would be installed as modules.

One solution to this would obviously be to use require() and
require().resolve() for this. This may still well be the best solution, but
given how the code is currently organized, it seems this would require
significant surgery.

Instead I change the behavior of jsdoc/path.getResourcePath() here in the
following ways, guided by making it _more_ consistent with what require()
would do so there are fewer surprises:

1. Distinguish between file paths relative to the module or working directory,
   and those relative to where installed packages are found. (Absolute paths
   left as is, as previously.)
2. If relative to module or working directory, look first there, then in
   the directory of the config file (this previously came first - I can
   hardly think of a reason why that would be logical), and only then as
   a fallback in the directory where JSDoc is installed.
3. If not starting with './' or '../', treat as relative to where installed
   packages are found. Look first under `node_modules/` in the current
   module/package and JSDoc configuration file directories, and then in the
   directory where JSDoc is installed.
4. For file paths starting with 'node_modules/' due to legacy instructions
   from 3rd party plugin and template authors, prefix with './' and then
   treat as relative to module/working directory (2,)

I also added tests that would pin this behavior down for the future. All
previously existing tests pass, despite the changes in resolution behavior;
this could obviously be because of gaps in test coverage, but it also suggests
that nothing truly relied on the brokenness of the previous behavior.

Fixes #1081.

[comment]: https://github.com/jsdoc3/jsdoc/commit/6be9dac616549ce226e4c0064876c7cb3b64fe25#commitcomment-18544292