Fixes detection of lack of directory in plugin def
This change removes the assumption that plugin specifications use the
path separator native to the system - should now work for whatever way
works for path.dirname() and path.basename().

The situation in which this won't make the correct determination is if
an installed package starts with a dot, and is given without directory.
A plugin or template for JSDoc that installs as a package and whose name
starts with a dot seems like an outer fringe case though that I'm not sure
is worth worrying about.