Deferred process.exit() to allow STDOUT pipe to flush
Since process.exit() is an imperative that exits immediately, pending
contents written to STDOUT is lost, if STDOUT is connected to a pipe
(as it is, for example, when used by sphinx-js to generate automatic
class descriptions).

The deferred exit in this commit allows STDOUT to flush its full
contents, before the jsdoc process exits with the designated exit code.

Please see
https://github.com/nodejs/node-v0.x-archive/issues/3737#issuecomment-19156072
for details.