fix(jsdoc): use the correct scope for exported classes with constructors
Previously, if a module exported an ES2015 class, and the class had a constructor, the generated docs claimed that the class was in the `inner` scope. But `inner` means that the class is _not_ exported.

This issue occurred because we generate separate doclets for the class declaration and constructor. When we merged the two, we incorrectly gave priority to properties in the constructor's doclet, rather than the class declaration's doclet. The constructor's doclet uses the `inner` scope; the class declaration's doclet uses the `static` scope.