fix: prevent circular refs when params have the same type expression
Catharsis caches parse results by default; if you parse the same type expression twice, with the same options, you get the same object each time.

When the user passes the `--debug` flag, we expose the parsed type for each parameter as an enumerable property of the doclet. If two parameters used the same type expression, the resulting doclet could contain a circular reference.

This change disables the Catharsis cache, so that Catharsis returns a new object for each parsed type, which prevents circular references. As a result, this change fixes an issue with the `-X` flag, and with some JSDoc templates.