Skip over broken files when scanning the source
This fixes the bug #1623 and other similar issues.

When scanning source code, the previous behaviour was to trip up on any
issue that `statSync` may raise. This can particularly happen for
circular and broken symlinks.

The fix addresses this by creating a mock version of `fs` that does not
even return such broken files from `readdirSync`.

The alternative fix would be to raise this on the `klaw-sync` package,
but that might be a breaking change for the packages that depend on it.