Deal with HTML paragraphs, without an ending period
It is pretty common for the markdown plugin to generate description
where the first line is wrapped in <p>. There is already an existing
line to deal with and remove these <p> explicitely.

This was however not working when the first line didn't end with a .
The split caused `<p>summary</p>.` combinations because we force
added a . after the non-matching split. This combination then
subsequently did not match the <p> stripping regex at the end.

By force splitting on end of line we can avoid this mistake.